+++ steinerpy/library/misc/utils.py
@@ -213,6 +213,9 @@ class ChangingKeysDict(UserDict):
     """Class to keep track of changing keys which are 
         composed of base_keys
 
+        TODO: Need to handle merging and when data keys contain other
+        things besides base_keys
+
     """
     def __init__(self, base_keys: list, *args, **kwargs):
         super().__init__(*args, **kwargs)